SHIELD-137: improve package name & type parsing, plus new checks

+++ generate_shield_package.sh
@@ -7,24 +7,37 @@ error() {
   exit 0
 }
 
+[[ ! -f $p_path ]] && error "Could not find package in file system"
+
 p_name=$(basename $p_path)
-p_id="${p_name%.tar.gz}"
+#p_id="${p_name%.tar.gz}"
+p_id=$p_name
 
 [[ -z $p_id  || -z $p_path ]] && error "The path to an OSM package must be provided"
 
-p_type="${p_id##*_}"
-p_type="${p_type%.*}"
-p_id_s="${p_id%_*}"
+# Apply regexp and capture name and type for package
+if [[ "$p_id" =~ ([a-zA-Z0-9]{1,})_([a-zA-Z0-9]{1,})[_(?:.*)]* ]]; then
+  p_id_s=${BASH_REMATCH[1]}
+  p_type=${BASH_REMATCH[2]}
+  # Replace OSM-specific naming
+  p_type="${p_type/vnfd/vnf}"
+  p_type="${p_type/nsd/ns}"
+fi
 
 [[ -z $p_id_s ]] && error "Could not determine name of the package"
 [[ -z $p_type ]] && error "Could not determine package type (NS or VNF). Expected formats: \${pkg_name}_ns.tar.gz, \${pkg_name}_vnf.tar.gz"
 
 repo_root=$PWD
+
+manifest_path="${repo_root}/security-manifest/${p_type}/${p_id_s}/manifest.yaml"
+echo $manifest_path
+[[ ! -f $manifest_path ]] && error "Could not find SHIELD security manifest for this package"
+
 pkg_tmp=$(mktemp -d)
 
 # Copy package and corresponding security manifest
 cp -Rp $p_path $pkg_tmp/
-cp -p security-manifest/${p_type}/${p_id_s}/manifest.yaml $pkg_tmp/
+cp -p $manifest_path $pkg_tmp/
 cd $pkg_tmp
 
 juju_in=$(dpkg -l | grep juju)
